+++ sdk/android/r2rpc-sdk/src/main/kotlin/io/r2rpc/sdk/Models.kt
@@ -1,4 +1,4 @@
-package io.rer0rpc.sdk
+package io.r2rpc.sdk
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable